requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
curlify==2.2.1
Django==5.0.3
djangorestframework==3.15.1
hyundai_kia_connect_api==3.17.9
idna==3.6
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.2.1
